refactor: improve architectural layering and resolve circular dependencies
- Extract utility functions to separate modules (date_utils.py, amount_utils.py)
- Move database operations from InvoiceController to Invoice model layer
- Add CustomerController.get_customer_by_name() wrapper method
- Move generate_invoice_metadata_from_filename to InvoiceController
- Update all imports to use controller layer instead of direct model calls
- Maintain clean separation: CLI → Controllers → Models
- All tests passing, 10.00/10 pylint score maintained

Addresses PR review feedback about circular dependencies and controller
database operations while improving overall architectural consistency.
